refactor: collapse peerSystem proxy to describePeer; tighten peerSection signature
Post-audit cleanups from a re-read of what shipped earlier today:

- `peerSystem(sync, id).describe()` proxy collapsed to a top-level
  `describePeer(sync, id)` function. The proxy shape was overkill for one
  method and the only justification ("future extensibility for system.list")
  is the YAGNI smell the spec itself called out as out-of-scope. Lives in
  `peer.ts` next to `peer<T>()` — co-located, same race-handling pattern.
- `peerSection`'s `if (!sync)` defensive branch was dead at runtime: the
  --peer path can't reach it (waitForPeer needs sync to find anyone) and
  the --all path skips it (empty peers map when no sync). Tightened the
  signature to non-nullable `SyncAttachment` and dropped the dead branch
  + its tautological test.
- `openNotes({ device })` → `openNotes(device)`. Destructure-for-future-args
  is the same imaginary-future smell.
- Comment in `notes.ts` explaining arktype's `_v: '1'` numeric-literal
  convention so a repro reader doesn't think it's a typo.

No behavior change. 636 pass / 0 fail. Local list smoke still works.
