+++ examples/notes-cross-peer/notes.ts
@@ -15,9 +15,12 @@ import * as Y from 'yjs';
 const SERVER_URL = 'https://api.epicenter.so';
 const WORKSPACE_ID = 'epicenter.notes-repro';
 
+// `_v: '1'` here is arktype syntax for the literal NUMBER 1 (numeric strings
+// in arktype's type position resolve to number literals). The `set()` call
+// below passes `_v: 1` — same value, two different syntax conventions.
 const Note = defineTable(type({ id: 'string', body: 'string', _v: '1' }));
 
-export function openNotes({ device }: { device: DeviceDescriptor }) {
+export function openNotes(device: DeviceDescriptor) {
 	const ydoc = new Y.Doc({ guid: WORKSPACE_ID });
 	const tables = attachTables(ydoc, { notes: Note });
