+++ examples/notes-cross-peer/peer-a/epicenter.config.ts
@@ -1,5 +1,7 @@
 import { openNotes } from '../notes';
 
 export const notes = openNotes({
-	device: { id: 'notes-repro-peer-a', name: 'Peer A', platform: 'node' },
+	id: 'notes-repro-peer-a',
+	name: 'Peer A',
+	platform: 'node',
 });
