+++ examples/notes-cross-peer/peer-b/epicenter.config.ts
@@ -1,5 +1,7 @@
 import { openNotes } from '../notes';
 
 export const notes = openNotes({
-	device: { id: 'notes-repro-peer-b', name: 'Peer B', platform: 'node' },
+	id: 'notes-repro-peer-b',
+	name: 'Peer B',
+	platform: 'node',
 });
